Move slider to separate package
