+++ packages/core/package.json
@@ -49,7 +49,6 @@
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
     "react-test-renderer": "^16.5.2",
-    "recompose": "^0.30.0",
     "styled-components": "^3.4.9",
     "stylis": "^3.4.9",
     "svgi": "^1.1.0"
