+++ packages/core/src/index.js
@@ -31,7 +31,6 @@ export { default as RedButton } from './RedButton'
 export { default as Relative } from './Relative'
 export { default as Stamp } from './Stamp'
 export { default as Select } from './Select'
-export { default as Slider } from './Slider'
 export { default as Text } from './Text'
 export { default as Tooltip } from './Tooltip'
 export { default as ToggleBadge } from './ToggleBadge'
