+++ packages/slider/README.md
@@ -1,0 +1,27 @@
+
+# pcln-slider
+
+React range slider component
+
+```sh
+npm i pcln-slider
+```
+
+```jsx
+import Slider from 'pcln-slider'
+```
+
+```jsx
+<Slider
+  min={0}
+  max={0}
+  value={value}
+  onChange={handleChange}
+/>
+```
+
+## Props
+
+Uses the same props as [rc-slider][].
+
+[rc-slider]: https://github.com/react-component/slider
