+++ packages/slider/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  presets: ['@babel/env', '@babel/react']
+}
