+++ packages/slider/test/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import { Slider } from '..'
+import Slider from '../src'
 
 describe('Slider', () => {
   test('renders', () => {
